+++ demo/app/app.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit, ViewEncapsulation } from '@angular/core';
+import { AfterViewInit, Component, OnInit, ViewEncapsulation } from '@angular/core';
 import { WebSocketSupplier } from '../utils/services/webSocketSupplier/webSocketSupplier';
 import { MapLayerProviderOptions } from '../../src/models/map-layer-provider-options.enum';
 import { ViewerConfiguration } from '../../src/services/viewer-configuration/viewer-configuration.service';
@@ -6,6 +6,7 @@ import { DomSanitizer } from '@angular/platform-browser';
 import { MdDialog, MdIconRegistry } from '@angular/material';
 import { AppSettingsService } from './services/app-settings-service/app-settings-service';
 import { ViewerFactory } from '../../src/services/viewer-factory/viewer-factory.service';
+import { ViewersManagerService } from '../../src/services/viewers-service/viewers-manager.service';
 
 @Component({
 	selector: 'app-root',
@@ -15,8 +16,9 @@ import { ViewerFactory } from '../../src/services/viewer-factory/viewer-factory.
 	encapsulation: ViewEncapsulation.None
 })
 
-export class AppComponent implements OnInit{
-	arcGisMapServerProvider = MapLayerProviderOptions.ArcGisMapServer;
+export class AppComponent implements AfterViewInit{
+  
+  arcGisMapServerProvider = MapLayerProviderOptions.ArcGisMapServer;
 	flyToOptions = {
 		duration: 2,
 		destination: Cesium.Cartesian3.fromDegrees(-117.16, 32.71, 15000.0),
@@ -27,7 +29,7 @@ export class AppComponent implements OnInit{
 							iconRegistry: MdIconRegistry,
 							sanitizer: DomSanitizer,
 							private dialog: MdDialog,
-							private viewerFactory: ViewerFactory) {
+							private viewersManager: ViewersManagerService) {
 		iconRegistry.addSvgIcon(
 			'settings',
 			sanitizer.bypassSecurityTrustResourceUrl('/assets/settings.svg'));
@@ -56,9 +58,9 @@ export class AppComponent implements OnInit{
 		sidenav.open();
 	}
 	
-	ngOnInit(){
-		// example for getting the viewer outside of the ac-map hierarchy
-		const viewer = this.viewerFactory.getViewer('map-1');
-	}
+  ngAfterViewInit(): void {
+    // example for getting the viewer by Id outside of the ac-map hierarchy
+    const viewer = this.viewersManager.getViewer('main-map');
+  }
 
 }